refactor(contacts): decompose contacts page into 4 focused sub-components
- Created src/components/contacts/ with 4 components:
  - ContactFormModal.tsx (368L) - add/edit contact modal
  - ContactList.tsx (294L) - list display with search
  - ContactDeleteDialog.tsx (216L) - delete confirmation
  - ContactFilters.tsx (203L) - filtering controls
- Reduced src/pages/contacts.tsx from 1,139L to 192L (thin shell)
- All 37 characterization tests passing
- Largest component: 368L (< 400L limit)
